src/requirements.txt
easydict
numpy
scipy
h5py
Pillow
tqdm
Cython
scikit-learn